Improve module boundary/privacy
