+++ src/config/deployment_cfg_types.rs
@@ -11,14 +11,14 @@ from_env_var!(
     /// The current environment, which controls what file to read other ENV vars from 
     let name = Env;
     let default: EnvInner = EnvInner::Development;
-    let (env_var, allowed_values) = ("RUST_ENV",  format!("one of: {:?}", EnvInner::variants()));
+    let (env_var, allowed_values) = ("RUST_ENV",  &format!("one of: {:?}", EnvInner::variants()));
     let from_str = |s| EnvInner::from_str(s).ok();
 );
 from_env_var!(
     /// The address to run Flodgatt on
     let name = FlodgattAddr;
     let default: IpAddr = IpAddr::V4("127.0.0.1".parse().expect("hardcoded"));
-    let (env_var, allowed_values) = ("BIND", "a valid address (e.g., 127.0.0.1)".to_string());
+    let (env_var, allowed_values) = ("BIND", "a valid address (e.g., 127.0.0.1)");
     let from_str = |s| match s {
         "localhost" => Some(IpAddr::V4(Ipv4Addr::LOCALHOST)),
         _ => s.parse().ok(),
@@ -28,35 +28,35 @@ from_env_var!(
     /// How verbosely Flodgatt should log messages
     let name = LogLevel;
     let default: LogLevelInner = LogLevelInner::Warn;
-    let (env_var, allowed_values) = ("RUST_LOG",  format!("one of: {:?}", LogLevelInner::variants())); 
+    let (env_var, allowed_values) = ("RUST_LOG",  &format!("one of: {:?}", LogLevelInner::variants())); 
     let from_str = |s| LogLevelInner::from_str(s).ok();
 );
 from_env_var!(
     /// A Unix Socket to use in place of a local address
     let name = Socket;
     let default: Option<String> = None;
-    let (env_var, allowed_values) = ("SOCKET", "any string".to_string());
+    let (env_var, allowed_values) = ("SOCKET", "any string");
     let from_str = |s| Some(Some(s.to_string()));
 );
 from_env_var!(
     /// The time between replies sent via WebSocket
     let name = WsInterval;
     let default: Duration = Duration::from_millis(100);
-    let (env_var, allowed_values) = ("WS_FREQ",  "a number of milliseconds".to_string());
+    let (env_var, allowed_values) = ("WS_FREQ",  "a number of milliseconds");
     let from_str = |s| s.parse().map(Duration::from_millis).ok();
 );
 from_env_var!(
     /// The time between replies sent via Server Sent Events
     let name = SseInterval;
     let default: Duration = Duration::from_millis(100);
-    let (env_var, allowed_values) = ("WS_FREQ", "a number of milliseconds".to_string());
+    let (env_var, allowed_values) = ("WS_FREQ", "a number of milliseconds");
     let from_str = |s| s.parse().map(Duration::from_millis).ok();
 );
 from_env_var!(
     /// The port to run Flodgatt on
     let name = Port;
     let default: u16 = 4000;
-    let (env_var, allowed_values) = ("PORT", "a number between 0 and 65535".to_string());
+    let (env_var, allowed_values) = ("PORT", "a number between 0 and 65535");
     let from_str = |s| s.parse().ok();
 );
 from_env_var!(
@@ -66,7 +66,7 @@ from_env_var!(
     /// (including otherwise public timelines).
     let name = WhitelistMode;
     let default: bool = false;
-    let (env_var, allowed_values) = ("WHITELIST_MODE", "true or false".to_string());
+    let (env_var, allowed_values) = ("WHITELIST_MODE", "true or false");
     let from_str = |s| s.parse().ok();
 );
 /// Permissions for Cross Origin Resource Sharing (CORS)
