+++ src/parse_client_request/mod.rs
@@ -1,7 +1,9 @@
-//! Parse the client request and return a (possibly authenticated) `User`
-pub mod query;
-pub mod sse;
-pub mod subscription;
-pub mod ws;
+//! Parse the client request and return a Subscription
+mod postgres;
+mod query;
+mod sse;
+mod subscription;
+mod ws;
 
-pub use subscription::{Stream, Timeline};
+pub use self::postgres::PgPool;
+pub use subscription::{Stream, Subscription, Timeline};
