+++ src/parse_client_request/sse.rs
@@ -1,78 +1,4 @@
 //! Filters for all the endpoints accessible for Server Sent Event updates
-use super::{
-    query::{self, Query},
-    subscription::{PgPool, Subscription},
-};
-use warp::{filters::BoxedFilter, path, Filter};
-#[allow(dead_code)]
-type TimelineUser = ((String, Subscription),);
-
-/// Helper macro to match on the first of any of the provided filters
-macro_rules! any_of {
-    ($filter:expr, $($other_filter:expr),*) => {
-        $filter$(.or($other_filter).unify())*.boxed()
-    };
-}
-
-macro_rules! parse_query {
-    (path => $start:tt $(/ $next:tt)*
-     endpoint => $endpoint:expr) => {
-        path!($start $(/ $next)*)
-            .and(query::Auth::to_filter())
-            .and(query::Media::to_filter())
-            .and(query::Hashtag::to_filter())
-            .and(query::List::to_filter())
-            .map(
-                |auth: query::Auth,
-                 media: query::Media,
-                 hashtag: query::Hashtag,
-                 list: query::List| {
-                    Query {
-                        access_token: auth.access_token,
-                        stream: $endpoint.to_string(),
-                        media: media.is_truthy(),
-                        hashtag: hashtag.tag,
-                        list: list.list,
-                    }
-                 },
-            )
-            .boxed()
-    };
-}
-pub fn extract_user_or_reject(
-    pg_pool: PgPool,
-    whitelist_mode: bool,
-) -> BoxedFilter<(Subscription,)> {
-    any_of!(
-        parse_query!(
-            path => "api" / "v1" / "streaming" / "user" / "notification"
-            endpoint => "user:notification" ),
-        parse_query!(
-            path => "api" / "v1" / "streaming" / "user"
-            endpoint => "user"),
-        parse_query!(
-            path => "api" / "v1" / "streaming" / "public" / "local"
-            endpoint => "public:local"),
-        parse_query!(
-            path => "api" / "v1" / "streaming" / "public"
-            endpoint => "public"),
-        parse_query!(
-            path => "api" / "v1" / "streaming" / "direct"
-            endpoint => "direct"),
-        parse_query!(path => "api" / "v1" / "streaming" / "hashtag" / "local"
-                     endpoint => "hashtag:local"),
-        parse_query!(path => "api" / "v1" / "streaming" / "hashtag"
-                     endpoint => "hashtag"),
-        parse_query!(path => "api" / "v1" / "streaming" / "list"
-                endpoint => "list")
-    )
-    // because SSE requests place their `access_token` in the header instead of in a query
-    // parameter, we need to update our Query if the header has a token
-    .and(query::OptionalAccessToken::from_sse_header())
-    .and_then(Query::update_access_token)
-    .and_then(move |q| Subscription::from_query(q, pg_pool.clone(), whitelist_mode))
-    .boxed()
-}
 
 // #[cfg(test)]
 // mod test {
